backend/requirements.txt
beautifulsoup4==4.4.1
cssselect==0.9.1
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.0
jieba3k==0.35.1
lxml==3.5.0
newspaper3k==0.1.7
nltk==3.2
Pillow==3.1.1
python-dateutil==2.5.0
PyYAML==3.11
requests==2.9.1
six==1.10.0
tldextract==1.7.5
urllib3==1.14
